requirements.txt
blinker==1.9.0
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
docxtpl==0.20.2
et_xmlfile==2.0.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
openpyxl==3.1.2
packaging==26.0
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pypdfium2==5.0.0
python-docx==1.2.0
python-dotenv==1.2.1
reportlab==4.4.5
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.0.1
